Extract proc macro into own crate
This helps testability of the proc macro, allows fuzzing it to find
panics, and would also allow benchmarking.
