+++ macros/macros_impl/src/encode.rs
