+++ macros/macros_impl/src/enum.rs
